requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
alembic
pydantic
openai==0.28
sentence-transformers
pinecone
jinja2
python-multipart
